refactor:move CategoryTraceFilter to InfluxDB.core
